IGNORE previous commit. I have some problems now:
unil now we were always storing info_type into information. but really
information needs the specific decorator because when editing an
information you need to know if it is required etc.
MY METHOD FOR GETTING THE DECORATOR OF AN INFO_TYPE VIA THE ACTOR TYPE
DOESN'T WORK AND I DONT KNOW WHY.